refactoing and tiebreak
