+++ src/lib/election-system/index.ts
@@ -1,4 +1,5 @@
+import HareClark from "./hare-clark";
 import Race from "./race";
 import Vote from "./vote";
 
-export { Vote, Race };
+export { Vote, Race, HareClark };
